requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Hugging Face Hub
huggingface_hub>=0.19.0

# Audio Processing (for Whisper/TTS)
librosa>=0.10.0
soundfile>=0.12.0
TTS>=0.19.0

# API and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Evaluation and Metrics
evaluate>=0.4.0
sacrebleu>=2.3.0
rouge-score>=0.1.2

# Development Tools
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Multi-modal
opencv-python>=4.8.0
